pyproject.toml
[project]
name = "money-manager-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-filter>=25.1",
    "faker>=36.1.1",
    "graphene-django>=3.2.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.9.9",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
    "pytest>=8.3.5",
]
